requirements.txt
APScheduler
django-cors-headers
django-rest-authtoken